functions/update_stock_data/requirements.txt
functions-framework>=3.0
yfinance>=0.2.31
google-cloud-bigquery>=3.0
google-cloud-storage>=2.0
